bin/upload-docs: set content type on extensionless files
We strip file extensions when rendering static files in order to
provide for pretty, extension-free URLs. Unfortunately, this
removes S3's ability to deduce the MIME type from the extension.

This commit adds a post-sync step that iterates over every
extensionless file in the generated docs, and updates the content
type to text/html. For now, we're not concerned about false
positives.

One other option we explored was having md2html render non-index
leaf pages as <basename>/index.html, but this caused relative URLs
in the content to be off by one level of depth. Changing relative
URLs to root-prefixed URLs (/foo/bar) was a deep change, and broke
the ability to browse the Markdown links from GitHub.